requirements.txt
demucs
spotipy
python-dotenv
spotdl
fastapi[standard]
jinja2
mutagen